requirements.txt

Flask==1.1.2
google-cloud-pubsub==1.6.1
google-cloud-bigquery==1.27.2

